requirements.txt
panda3d==1.10.16
panda3d-gltf
panda3d-simplepbr